refactor(js): migrate paywall html to shared folder
