+++ packages/typescript/x402/src/shared/paywall.ts
